+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - Set `swiftDialogMinimumRequiredVersion` to `2.5.0.4761`
 - Improved exit code processing for 'Welcome' dialog
 - Added pre-flight check for AC power (thanks for the suggestion, @arnoldtaw!)
-
+- Added Variables for Prefill Email and Computer Name (thanks, @AndrewMBarnett!)
 
 ## 1.14.2
 ### 15-Feb-2024